add infection
When infected, human will change to zombie within one day.

+++ zombie.c
@@ -44,8 +44,8 @@ void unlock(int i, bool *locks)
 
 int main(int argc, char **argv)
 {
-    srandom(time(NULL));
-    srand48(time(NULL));
+    srandom((unsigned int)time(NULL));
+    srand48((unsigned int)time(NULL));
 
     bool *locks = (bool *)malloc((SIZEX + 2) * sizeof(bool));
